Unpin package versions from Dockerfile
Pinned package version makes it a pain to rebuild the image when I check
in with this project.

Check <https://github.com/hadolint/hadolint/issues/884> for a rationale
on ignoring the 'pin package version' rule in apt-get installs.

Also did away with the `apt-get clean` as it seems that official Debian
images run it automatically
(see <https://docs.docker.com/develop/develop-images/instructions/#apt-get>).
postgres is based on the Debian image, I checked.